ci: Don't vote Verified-1 on soft-failed builds
Updatee the set_gerrit_status script to properly handle steps that are
marked soft_fail in buildkite, voting Verified-1 only if there are
failed jobs in the build that *aren't* soft-fail.

Since the buildkite GraphQL API doesn't allow filtering jobs by
softFail, this required tweaking the query slightly to first load a list
of failed jobs alongside the soft_fail attribute of those jobs, then
filter down to the ones with soft_fail not equal to true in jq.

Change-Id: Ib452f9b600572ba8aec1646749dafce4d386dd90
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1538
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>