consensus: Only the current leader can update the controller state.
Previously there were two scenarios where leaders could update the
controller state after consul viewed them as dead. (1) If a
controller began updating the controller state and was declared
dead before finishing the update, (2) If a network partition
is repaired and a node thinks itself the leader of the cluster
while the session is considered dead on split-brain recovery.

This CL updates the consul authority to require the leader's
session lock be used to lock the controller state. This is
done only if the server is the current leader at the start of the
call. If the session is still alive when the PUT occurs, then
the node must still be the leader when it updates the controller state,
otherwise the PUT operation will fail with an error indicating
that the session failed.

A caveat of this is that the authority no longer *deletes* keys
on session removal, otherwise the controller state would be
deleted, instead, they only release the lock. This has implications
for worker/adapter keys that may churn, and should be fixed at a
future time.

Change-Id: Ief1d58193b611da32302b721e52b94bf4fb6c5a8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1545
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>