Improved Eviction Benchmark
Rework the eviction benchmark to only execute the required
query, without any phantom background queries. Instead the
benchmark is capable of scaling down the query spec in order
to hit the desired cache hit ratio. In order for it to work
properly the query range in query spec should be sufficiently
large to be able to cause evictions at a lower hit rate because
the benchmark is only capable to scale down, not up.

Change-Id: Ie024a03c6fd1c29c93cca1a0976062572d74d915
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1581
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>