logictests: Fix postgres logictest framework + run in nightly soft fail
Logictest framework previously tried to prepare and execute
statements issued via `query_drop`. This causes us to throw
unsupported errors in the adapter, as we do not support
preparing certain DDL statements, CREATE TABLE in this case.

This replaces the `execute` call to `simple_query` for
`query_drop` and enables the logictests in nightly with soft fail
enabled. At this time 258/307 logictests failed.

Change-Id: Id6c7bbcdb0e7e246e02ea058a2f3e97ef291d877
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1555
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>