fault tolerance: Drop packets to failed domains.
Currently, when a domain attempts to send packets to a domain
that has failed, or if the packet send operation returns any
error, we fail the domain. This is the cause of cascading
domain failures.

This CL opts to instead drop *all* packets to failed domains,
forever. A follow-up change will inform the leader that
the domain is suspected to be failed to repair the dataflow graph
in the case the domain did not truly fail.

Change-Id: Iff95dfc0c9288177ddfa66369f50a915fd199c97
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1606
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>