ops/buildkite: add ssh key to CF and allow private subnet agents
Adding an SshKey parameter option to the buildkite queue module so that
we can ssh into the boxes.

After reviewing the CF templates, it appears that a blank value for the
SSH key should cause no changes to the other/existing queues.

When a KeyPair is passed to the CF template, a 0.0.0.0/0:22 rule is
automatically added to the VPC security group assigned to the agent EC2.

Given that, I've added support for running agents in private subnets; it
was fairly straight forward to add in. Since our VPCs have a NAT
gateway, this shouldn't be an issue. I figured we can try things out on
the  benchmark queue first, and then move the other queues over when it
makes sense. Effectively, we're just mitigating some risk with this
change by limiting it.

Additionally, I've increased the upper bound of the main queue (to 20).

Change-Id: I2e64abc3788abc1efc18e1b42b5b4f3ce499cc13
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1558
Reviewed-by: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI