unbounded-interval-tree: Take more things by-reference
In order to get fully functional partial range semantics working
correctly in reader_map, I need to replace the backing store of the
BTree version with PartialMap - however, in order to make the type
signature work out for that, things in unbounded-interval-tree need to
be slightly more flexible. Specifically, the `get` method for reader_map
takes `&Q where K: Borrow<Q>` similarly to std's HashMap, allowing in
our case passing `&[T]` for a map with a `Vec<T>` key - but eg
`IntervalTree::contains_point` previously only took `&K` - which as a
side-note it then proceeded to *clone*, which is no good! Making this
more flexible took me down a bit of a rabbit hole taking and returning
more things by-reference in unbounded-interval-tree, and avoiding
cloning and allocation in quite a few places - perhaps most notably, the
`contains_interval` method now avoids allocating a whole vector of the
interval difference just to check if it's empty, instead inlining most
of the logic from `get_interval_difference` and short-circuit returning
a boolean if we'd otherwise append onto the vector of the difference.

This whole commit doesn't really change anything functionally, but
paves the way for fixing partial range coverage in the reader, and
should improve performance rather significantly in a few cases!

Change-Id: I2f0ae75d9c2b86a95e1211bf9cac158dc511faff
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1568
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>