ops/substrate: implement sudo addtl perms for buildkite tf module
The purpose of this CL is to allow one to pass in supplemental Sudo
permissions to the underlying Buildkite agent stack.

For the purposes of benchmarks, I've added the NOPASSWD:ALL permission,
as indicated in the following issue:

https://bit.ly/3qRQX6I

In order to mount files into the benchmark MySQL container in CI, we're
going to need to be able to change the file permissions outside of the
container first. By default, the buildkite agent prevents you from using
root. To that end, benchmarking needs these changes to work.

@Grfn mentioned that we're going to need this elsewhere as well, so this
should be a general win for us.

Also took a quick moment to separate `data` resources from tf `resource`
objects. This is a common approach to keep code digestible as its'
footprint increases.

Change-Id: I429e9851b3a9ee024ff5ae286a75710c1c1b1e5d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1565
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>