fault tolerance: fail worker on domain failure.
This CL causes the worker to panic instead of silently fail on
domain failure. This will increase the blast radius of a domain
failure to the entire worker (several domains and potentially
the leader), but we expect domain failure to only handle in extremely
rare failure scenarios likely caused by bugs or data corruption on
the worker's disk.

Updates two clustertests that now pass as we recover from the
failure due to worker failure.

Change-Id: I8990c2edb8b0b1f2f4a79cb672b717064426918d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1607
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>