reader_map: Replace backing store for btree with PartialMap
Previously, in btree mode, the reader_map didn't properly respect ranges
when doing point lookups or inserts - meaning it'd return a Miss if you
ever tried to retrieve a point that was covered by a range that was
marked filled but didn't actually have any keys. This would, among other
things, cause writes going to a reader_map were covered by filled ranges
to incorrectly get dropped.

Fortunately, we already have a data structure that encapsulates both a
BTreeMap and an interval tree with filled ranges - the PartialMap that
was previously in noria_dataflow::state. This commit extracts the
PartialMap structure out into its own crate so that reader_map can
depend on it, then updates the backing storage for the BTree version of
reader_map to use a PartialMap instead of storing its own btree and
interval tree directly. This gets us the correct semantics for lookups
of points covered by filled ranges for free.

All of this comes with a previously-failing logictest covering the
original issue, and a couple of extra trace logs I added to the write
path for readers that helped me debug the issue itself.

Change-Id: I9f951b189181629c91eb984d00a7767c7ec84508
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1579
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>