Fix TopK group_by bug
A TopK bug was previously uncovered involving compound-keyed topk. It
turned out that this only induced the bug if the WHERE AND clause was in
a different order than the column order that was established by the DDL.

It turned out that this was caused by sorting group_by when constructing
the TopK node. By doing so it caused a miss on the first replay due to
the group by not following the ordering of the replay. This set off a
chain of events that caused 2 more replays to get issued.

By the last replay we have two current records, one with a is_new of
false, and the other with an is_new of true. Because k is 1, we skip
over the first record and only end up seeing the one that has is_new
false, which leads to us not emitting any deltas, and therefore emitting
an empty result set when we should have returned 1.

By not sorting the group_by during TopK node construction, we get a hit
on initial state lookup, and emit the correct delta from TopK.

Change-Id: I4a78c8111bcfc1969aa99dbbd9f33e00131cb0b3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1582
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI