server: Don't drop subranges that *don't* miss
When we do a lookup of a range and only part of that range misses, we
only trigger a replay for the part of the range that missed - then keep
doing a lookup of that new range until it doesn't miss anymore (meaning
the replay completed). For the part of the range that *didn't* miss,
however, we totally forgot about it - meaning any results covered by
that part of the range would never actually get returned!

This wasn't an issue for point queries, since those either all miss or
all return results, and the miss / not miss bookkeeping was
per-KeyComparison.

There are a lot of ways to approach fixing this, but I decided to go
with what I thought was initially the simplest - just subtract the miss
from the original lookup key and re-do the lookup with the bit that
didn't miss, then store that in our buffer of results to eventually
return. This may look asymptotically poor at first glance, but keep in
mind that we currently only ever do lookups with one range at a time, so
n will usually be quite small, meaning the cost of building up a data
structure to do this subtraction is probably much more expensive than
the cost of iterating over some intervals and doing some light
subtraction. But if profiling shows this is slow down the road we can
always optimize, potentially by doing the difference and intersection
with the tree of ranges covered by the reader in one pass.

Regardless, this comes with a now-passing originally failing logictest,
and some extra trace logs I added to help debug this issue.

Change-Id: I982cf2e4bedc65fbc0864c6a0667b38c799cbc32
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1583
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>