noria-client: Don't autoparam queries with range params
We don't *yet* support mixing range parameters and non-range parameters
in the same query. We don't want the autoparametrization pass in
noria-client to take a query that *would* be supported and make it
unsupported, so add an extra pass before we run autoparametrization to
check if the query already has range params and skip autoparametrization
entirely if we find any.

We *could* make this clever enough to figure out if we can
autoparametrize all ranges or all equal filters, but we're going to be
adding support for mixed range and equal params soon enough that this
should probably just be kept simple for now.

Change-Id: Ie1d9a7c5b977f876ac62540b7204073fc4f48e56
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1592
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>