installer: Statically link libssl on linux-musl
We want to build a fully statically linked binary for the installer, so
that different glibc versions and available dynamic libraries don't
cause it to fail to run on the target machine. We want this to include
statically linking openssl, which the openssl crate allows us to do by
setting the "vendored" feature flag. We don't want to *always* do this
for openssl, however, so this CL just passes vendored for the linux-musl
target of the installer (aka when building a statically linked binary of
the installer). This along with setting resolver = "2" gets us static
openssl for the installer binary but a dynamic openssl everywhere else.

Change-Id: I0aae8f9c34dbece59ea8b150a03a8db303e6ab7a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1598
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI