launchpad: Factor out common property tests for trait laws
Factor out a few macros into a new `launchpad::properties` module for
defining proptests to check the laws of a few builtin traits, notably in
this case Hash, Eq, and Ord. This makes it easier to check these laws
when we're defining custom (non-derived) implementations of these
traits.

There were a few places I could find where we were manually defining
tests for these traits in the workspace, and I've replaced all of those
with invocations of these macros, but it's possible I've missed
something.

Change-Id: I0a4303b56b4fc49a8f12b3e58311d6dc5976ad32
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1608
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>