noria-data: Make Ord impl a proper total order
When defining custom comparators for rocksdb, it's important that the
comparator provided implements a proper total ordering relationship on
keys. If it doesn't, rocksdb gets *very* confused - the symptom I saw
was a "Data Corruption" error that said something about "two consecutive
TryAgain in WriteBatch handler", which stepping through GDB showed was
caused by the underlying MemTable implementation thinking it was seeing
the same key over and over again. Digging more into the data showed this
happened at the very least any time we tried to insert both None and
an Int into the same BTreeMap-backed column, since previously None > Int
and also Int > None, but lots of other different flavors of DataType
could've caused this.

This commit fixes the Ord impl for data type to get as close as possible
to a proper total ordering relation, driven by property testing via the
new launchpad::properties macros - lots of this generally involves big
catchall pattern-matches in the cmp function to make sure we're
comparing the various enum variants properly in order. The only caveat
here is that Numeric can't always be converted to an f32 or f64, and
vice versa, meaning that comparisons with Numeric can break transitivity
of both `>` and `<` - for now this is *probably* not a huge problem,
since we only ever insert things into state impls that are either
homogenously typed or None, but this might become a problem in the
future if that ever changes. To get that working with the property
tests, this also extends all the macros in launchpad::properties to take
an optional proptest strategy annotation for generating the values.

Change-Id: I8076940ba6f34d901093646df94127d68b5befad
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1609
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>