domains: drop packets to nodes that do not exist.
Migrations may run concurrently with packets being propagated through
the dataflow graph. As a result, a packet may be destined for a node
that does not exist. This is exceedingly rare, but we would not
want it to crash the domain/worker when it does occur. Instead log
an error.

Change-Id: I07921c9d3582bff75070b37760df0e5c53bd82b6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1611
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>