installer: exec binary, and don't redirect tty
I have absolutely no clue why, but doing things like this causes the
binary to stack overflow when run inside the wrapper script on Darwin.
In addition, not `exec`ing the binary causes it to fail with a "too many
open files" error. Again, not sure why, but it is what it is. Without
that we lose the automated cleanup after the fact, but it's all in a
tmpdir anyway so the OS should just clean it up for us regardless.

Change-Id: I97117715c1c3bc87317422bc5bc29fb8f59406dd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1613
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>