Refactor noria-client to remove clones on the critical path
Currently the statement execution path pays a big price
for performing lots of validation on that path, however
they can all be avoided and cached. The result is however
is a huge refactor (some of it not strictly needed) that
results in a cleaner implementation as a whole.

As a consequence the read throughput for the mysql adapter
is now 7% higher.

Change-Id: I9cf6c468e0e0d5c43daf85d70f06999c26e407a0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1656
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>