cfn: Update log reader to be a deploy user
This updates this to be a deploy user instead of just a log reader. This
should give us all the permissions we need. We can lock it down further
with limiting the IAM role that CloudFormation and other limitations but
IRL seems open to this and so let's suggest this and let them come up
with objections.

Change-Id: I70811db0374eab3404baf009e9da279a9d5ac531
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1619
Tested-by: Buildkite CI
Reviewed-by: Marcus Bastian <marcus@readyset.io>