ci: Don't rebuild images if dependent files are unchanged
The contents of several docker images in our pipeline only depend on the
Dockerfile itself and, in some cases the rust-toolchain file - but
notably not any of the code in the repository. Even with docker caches,
pulling building tagging and pushing this image takes around a minute
for each step if the agent we're running on is brand new.  If the
dependent files are unchanged, then, this commit just uses the AWS API
to re-tag the image using the current commit, which should ideally be
significantly faster (on the order of up to a couple minutes in the
worst case).

Change-Id: Icfa4dd7adcc6197b952c057ad253fa53391cb8dd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1624
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>