adapter: Support for returning a table of metadata.
This introduces a new QueryResult variant: MetaVariables. Which
are used to represent a table of variables similar to those returned
by SHOW STATUS in MySQL.

This will allow returning metadata with multiple rows in the following
format.
+--------------------------+------------=+
| Variable_name            | Value       |
+--------------------------+-------------+
| replication_status       | IN_PROGRESS |
| allowed_queries          | 4           |
| proxied_queries          | 14          |
| cache_utilization        | 40%         |
+--------------------------+-------------+

The existing way to return metadata, [`QueryResult::Meta`], instead
mapped a key-value pair of metadata to a single row with a single
column. That fits a separate use-case where only a single value is
ever returned.

This new result type will be used to implement SHOW READYSET STATUS
in a follow up CL.

Change-Id: I2ce78933aafc331f8ea3e6ca6e529ddda52a439e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1629
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>