cfn: Create ReadySet CloudFormation Exec Role
In order to have a low access account that can go on to do higher levels
of access, you need to create what is called in various documentation an
Execution Role.

After using Athena to gather the list of APIs, then going through a full
cycle of creating and deleting the superstack, this appears to work in
our environment.

Change-Id: I2ba86b0cb750daf3a3cb142ddae3738268fdbad0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1625
Tested-by: Buildkite CI
Reviewed-by: Ryan Cerniglia <ryan@readyset.io>