ci: Fix external AMIs
Use the DeployCustomerArtifactsWrite policy from build to deploy to
write to the public CloudFormation bucket which already has a bucket
policy allowing this write.

Update external AMIs to use the release name for where it puts those
files in the public bucket.

Update locals to include the whole list of AMIs we expect demos for with
comments. Things under the new structure should be much more usable.

Change-Id: Id8d29289ba943bbebbb8658c8ebd04df785a4bcf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1626
Tested-by: Buildkite CI
Terraform-Plan-Review: Sasha Fox <sasha@readyset.io>
Reviewed-by: Ryan Cerniglia <ryan@readyset.io>
Reviewed-by: Marcus Bastian <marcus@readyset.io>