treewide: Properly disallow use of dbg!
The documentation[0] for the clippy dbg_macro lint explicitly states
that specifying it as a crate-level attribute doesn't work - the linked
clippy issue[1] has a comment suggesting instead enabling it using a
conditional rustflags cfg block in .cargo/config, which does appear to
work properly, as it caught one remaining dbg that accidentally landed
in main.

This commit adds that, and removes the crate-level
warn(clippy::dbg_macro) from all crates in the workspace, since it
doesn't do anything.

[0]: https://rust-lang.github.io/rust-clippy/master/#dbg_macro
[1]: https://github.com/rust-lang/rust-clippy/issues/6610

Change-Id: If9a3431bd605ba11f50e06b39d95a8d8eb5f09d8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1633
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>