server: Remove everything relating to ParamFilter
ParamFilter was a half-finished attempt at getting support for partial
LIKE and ILIKE in dataflow - it's no longer a thing that we're actively
working on as of many months ago, and to be perfectly honest I'm not
even convinced that the original design would have worked at all. For
now, it's clogging up the code and making maintenance a little bit more
difficult, so let's get rid of it.

Change-Id: Ibbd9a4617256e5684883e2c13efefeddbc0312b6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1638
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI