noria server: Don't panic on `inputs` if base tables is dropped.
The `inputs`, used to retrieve the set of base tables makes the
assertion that every node connected to the source node is a base table.
This does not hold when nodes are removed from the dataflow graph, which
can occur due to failures or migrations. When this happens the removed
nodes become `Dropped` nodes instead of `Base` nodes, failing the
assertion.

Updates to filter out dropped nodes when returning the inputs.

Change-Id: I49f29a9f8c8f34d97a09d3a43219af3446ad1b69
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1640
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>