server: Factor out generation of view keys
Refactor the generation of the lookup key columns and index for a view
into a method on QueryGraph, rather than doing it inline in
SqlToMirConverter::make_nodes_for_selection. This is a non-functional
change right now, but paves the way for adding a bit more complexity to
this conversion with the introduction of support for partial BETWEEN and
mixed range and equality comparisons.

Change-Id: I8bbd7d5c7c80052fc25ef55906870ab3fae0bba4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1639
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>