server: Add support for partial BETWEEN
When generating lookup keys for a query in the server, detect parameters
that look like a BETWEEN query (the same column with both a >= and a <=
for a column), and emit only a single range key column in the view for
that column. Then, carry the information that we just did that through
the `key_mapping` that's passed to the Leaf mir node, to the reader
node, and on to the view, via a new `ViewPlaceholder` enum which
replaces the previous `Option<PlaceholderIdx>` we had before.

Since parameter ordering very much matters now (and would have anyway,
to support mixed comparisons), and the way this is implemented causes
parameters to get output in a different order than before, this removes
the bit that was sorting parameter keys plus a test that was asserting
on the order of parameters - the `git blame` for the latter seems to
indicate that it was an *attempt* at covering a bug that someone had
found many years ago but didn't actually cover that bug, so it should be
fine to get rid of. Also, since `qg.parameters()` is no longer the
authoritative source of truth on the query's parameters, this updates
the generation of `group_by` columns for a topk to be based on the
columns in the `view_key` instead.

In noria-client, we currently match on the new ViewPlaceholder::Between
enum variant and return an unsupported error for it - next up, we'll
teach the adapter to look at this ViewPlaceholder enum and use that to
build a proper BETWEEN range key.

Change-Id: Ia5ab3aa82e204ebb4f2bb1a86140425af9017e70
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1647
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>