fault tolerance: alias readers by their query expression NodeIndex.
Query expressions are currently assigned the NodeIndex of the reader
node's parent in the *MirQuery*.

For example for a MirQuery:
[1] --> [2:Reader]
The query expression associated with the reader is assigned NodeIndex 1.

The dataflow graph might actually look like this:
[1] --> [3:Egress] --> [4:Ingress] --> [2:Reader]

When looking up reader nodes, we perform a BFS from [1] to find all the
readers. This does not work, however, when handling failures of a
server. For example if the server with nodes 2 and 4 failed, when we
try to determine what query expressions failed we are looking for node
1.

[1] --> [3:Egress] --> { [4:Ingress] --> [2:Reader] } Failed

This CL adds the NodeIndex of the query expression associated with the
reader node to the set of failed nodes.

Change-Id: I2b890611d0997f9bd416164bad14fa9641137723
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1644
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>