noria-client: Get parameter indices from the View
Rather than extracting key column indices in the adapter and carrying it
through the PreparedSelectStatement, directly use the new `key_map()`
exposed by View to get the parameter indices corresponding to the
individual key columns for a reader. This both gets us dropping
removed parameters eg when ORDER or LIMIT is removed by noria for
"free", allowing us to remove all that functionality from the
adapter, *and* makes us more robust in the future to things like
reordered key columns or key columns that don't correspond 1:1 to query
parameters.

We *do* still need to do a *little* bit of work in the adapter for limit
and offset parameters, namely to handle the edge case where the user
sets offset 0, or a limit of 1 but an offset other than 0, and
short-circuit to return an empty resultset - later, I'd like to move
this into the server if possible, potentially along with handling of
WHERE IN and other rewritten parameters.

Finally, we need to do a little extra work in the rewrite passes to get
placeholder numbering properly working in concert with both where-in
collapsing and autoparametrization, since both of those
make *positional* rather than *numbered* parameters, and since we now
actually care about parameter numbers it's important that when we give
those positional parameters numbers those numbers aren't already taken
by an existing user-supplied numbered parameter.

Change-Id: I5dc67c6b3573b64817a96703066a6b0dffe215f1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1650
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI