noria-data: Use EnumKind to compare DataType variants
Rather than the extremely manual, error-prone, and annoying way we were
previously comparing DataType variants, use the enum_kinds crate to
generate a zero-field `DataTypeKind` enum which *does* derive Ord, and
use *that* to compare not-usually-comparable DataTypes. I had
originally *attempted* to do this with std::mem::Discriminant, but that
doesn't derive Ord - fortunately, there's a crate for that! (tm).

The DataTypeKind enum we're generating here isn't exported yet,
primarily because I can't think of anything that would want to use it,
but if we find a different use for it later it'd probably be fine to
export.

Change-Id: I4be7bf90da363f8834bcee65ff2f0f2eb0a74cf6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1658
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>