readers: refactor readers to a stateful service
Refactors the reader `tower::Service` to be based off a struct
implementing `Service` instead of `service_fn`. This allows us to
maintain parameters for the entire service instead of passing them
through to the service_fn.

This will be used in a followup change to propagate an upquery timeout
to `BlockingRead`.

Change-Id: I4af6817b0a783ae288a18289c41b9451e9a3b7c4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1664
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>