timeouts: Support configuring upquery timeout in config
Upquery tasks may currently run forever waiting on the response from
upstream domains. In the case of failures, these tasks may never receive
a response to the upquery, for instance, if the message is being
dropped by an upstream domain. This timeout is currently set to
20 seconds to match the current View request timeout.

This change introduces an upquery timeout included in the `Config`.

Change-Id: I93564a28f543ebe00a284604e6faeb1c891bf910
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1665
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>