dataflow: Actually pass IndexType to reader_map
Actually thread through the index of a reader node to the initialization
of the reader_map when initializing readers, rather than just always
using a btree map. This involves updating the Global and PartialGlobal
variants of InitialState to have a full `Index` rather than just a `key:
Vec<usize`, the same way we did for PartialLocal and IndexedLocal.

I honestly can't believe I forgot to do this when I initially did index
selection - but with this change now that we're actually using hash maps
for IndexType::HashMap readers, apparently we have a 10-fold increase in
99%ile performance and a 16% increase in throughput (according to Vlad).

Change-Id: If2328d0118a99b4911061d31f16756042243e293
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1668
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>