noria-client: Reorder placeholders first
Numbered placeholders supplied by the user can come in any order, and
even be non-contiguous. This generally works fine now that we're pulling
parameters by index instead of in order, except in the case that
we *add* a parameter via the autoparametrization pass, in which case the
parameter numbering pass can give numbers to those new parameters that
overlap with numbers we already have.

We *could* do the extra work here of figuring out the maximum
user-supplied parameter and then shifting parameters around to add new
parameters after that, but it actually feels much simpler to just
rewrite the query such that the parameters are all in order, then
store the mapping between parameter indices in the ProcessedQueryParams
and use that to shuffle around the user-supplied set of parameters.

Change-Id: I065e31d55551cbfb41d6d448f8c592e5ea7fb3e5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1680
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>