timeouts: Configurable controller request/migration times.
This allows configuring the amount of time spent waiting on a
controller RPC. It splits controller RPC timeouts into two: one for
general requests, these typically query the controller for state and
complete quickly, one for migrations which will can take on the order
of minutes to tens of minutes for larger datasets.

Exposes these two as hidden flags, configurable in the adapter.

Change-Id: I18865d414e771b050f638e1b72ecedf938b351cc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1672
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>