timeouts: Support configuring worker request timeouts in config.
The worker provides a simple RPC interface similar to controller
requests. This CL replaces the existing RPC timeout stored in the
WORKER_RPC_REQUEST_TIMEOUT with a config entry.

The `do_noria_rpc` abstraction is also moved into `rpc` in this CL
as the abstraction up till now has not served a purpose.

Change-Id: I67559e7394e289ca7731d2341276f32c92576660
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1673
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>