timeouts: Set sensible timeouts
This CL updates all timeouts:
  * View RPCs (client queries): 5s
  * Table RPCs (replicator writes): 30 minutes
  * General controller RPCs: 5s
  * Migration RPCs: 30 minutes
  * Worker RPCs (used for migrations): 30 minutes

We expect client queries and general controller RPCs to be on the order
of hundreds of milliseconds *at worst*. Migrations may take an extremely
long time.

These RPCs can be re-evaluated and extended to be configurable via
command line flags if needed in the future.

Change-Id: I9c2ec8a02ec760a53f2c5a21af285f0983509a2f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1674
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>