benchmarks: Support writing benchmark outputs to file.
It will be useful to materialize the benchmark outputs in files that
can be used as artifacts in CI. This CL enables that through a new
command-line argument `--results-file`.

The results file is comprised of the yaml description of the benchmark
and deployment. It is created if it does not exist, and otherwise
appends the current benchmark's results.

Change-Id: Ia3a6de388ae2a24c1437b25f936357bc2c561297
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1676
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Marcus Bastian <marcus@readyset.io>