Update the metrics dependency
The recent version of the metrics crate made a big change
to how it works, by allowing one to get a handle to the
metric itself, as opposed to being limited to the macros.
https://github.com/metrics-rs/metrics/pull/240

For our Domain metrics the result should be a significantly
lower CPU usage, since currently the dynamic keys have to
be rebuilt every time which is an expensive operation.

The Domain changes will come in a follow up CL.

Change-Id: Iad5e571048138773af7dfb51cd77451f56df5587
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1681
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>