Send failed executes to fallback
This commit changes the adapter behavior by providing a new mechanism by
which was can have a recovery period in the case of a component outage
in a distributed deployment.

This is done on a per query level, where we record whether each
execution was successful, or not successful, and if not successful, if
it was networking related. This is recorded with a state transition
time.

We can then inspect this data to determine if a query has been
consistently failing for an unacceptable period of time, at which point
we enter a recovery period during which we send all queries to fallback.
After the recovery period we allow trying to query again.

This is done on a per query basis because in a distributed deployment
each query may rely on a different noria server. By assessing this on
the query level, we can have more fine grained control.

The benefit of such a feature is that during an outage we won't be
needlessly slamming a noria server when it's unreachable.

Change-Id: Ie4bf9240de098576dd2aa8835de6eaa7c5ec13e9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1689
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>